### 2019-08-29 #7 @PaulineBock
[Re] The Wisconsin Card Sorting Test: an implementation of Dehaene and Changeux Neuronal Network Model
**Original article:** Dehaene, S., & Changeux, J. P. (1991). The Wisconsin Card Sorting Test: Theoretical analysis and modeling in a neuronal network. Cerebral cortex, 1(1), 62-79.

**Scientific domain:** Computational Neuroscience
