### 2019-09-05 #7 @oliviaguest
@PaulineBock there is a minor typo in the penultimate line of the introduction, I think it should be "implementations" (plural).

@rougier why is there no checkbox for the type of replication?
