### 2019-09-05 #7 @PaulineBock
Hello Olivia Guest,
thank you for the review, and yes I think it should be plural, I fix it
right now.

Le jeu. 5 sept. 2019 à 18:41, Olivia Guest <notifications@github.com> a
écrit :

> @PaulineBock <https://github.com/PaulineBock> there is a minor typo in
> the penultimate line of the introduction, I think it should be
> "implementations" (plural).
>
> @rougier <https://github.com/rougier> why is there no checkbox for the
> type of replication?
>
