### 2019-09-09 #7 @AndreaCaso
Hello Pauline, here's my review.

**Code:**

In order to run the model on PyCharm I had to make just one (very) minor change. I added an argument to the main function (WCST_test in the file DehaeneChangeux.py) to prevent the output window from showing up, so that whenever one runs 500 trials in a row (in Statistics.py) PyCharm won't freeze up.
The code works very well, otherwise.

**Text:**

'to evaluate frontal lobe dysfunction': This could be rephrased a little, because a number of publications argue that the WCST in not as sensitive or specific to frontal lobe lesions as once thought.

'the subject is told to match the cards...': For a casual reader who is not familiar with the test, this is not very clear.  Perhaps citing Heaton (1981) could help direct the reader to the relevant literature

'processus': I believe this is a typo

'and a maximum value': what is this maximum value?

'without more precisions': could you please clarify?

'it is the reciprocal of speed of learning': why? could you please clarify?

In addition:

- Figure 3 shows the comparison between the original paper and the replication.
This is good, but the graph shows just one type of simulation (the implemented and modified network). What happened if one lesions the reward or the rule-coding network in the replicated model?

Also: I believe that the reasoning + memory condition is equivalent to the auto-evaluation + memory, is that correct? It would be helpful to the reader if these concept where clarified and consistent nomenclature was used.

- I understand that the modifications have been introduced in order to replicate the same results observed in the original paper because some details on the implementation weren't very clear (please tell me if I understand this correctly). If this is the case, I would also add a few lines in the Implementation and Modification paragraph and, most importantly, in the Conclusion, to describe the results obtained WITHOUT the modification and compare it against the one obtained WITH the modification.
What can be gleaned by all this in terms of cognitive processes?

I hope it helps. Feel free to ask if something is not clear.
Best,
Andrea
