### 2019-09-09 #7 @qihongl
Wow, this is really interesting! Thanks for the replication. I'm happy to help! I will start reading it next week!

I just looked at the code briefly. Just a quick minor comment: Making some object-oriented classes to separate the code for defining the model/task vs. the code for executing the experiment/plotting the results, might further enhance the readability and modularity of code :)
