### 2019-09-09 #7 @PaulineBock
Hello to all of you,

Thanks for the review. I will try to answer to all the questions.

I can rephrase the part about frontal lobe dysfunction and read more about
the clinical aims of the WCST.

I will cite Heaton and read more about it.

This is indeed a typo.

The maximum value depends of the long-term component values of the unit.

When I say that we have not more precisions, it is because on the original
paper, there is only one phrase about the activation of the go unit, where
the authors only says that it  is activated externally and that's all.
Maybe I could cite this phrase ?
"Propagation is gated by the activity of a "go" unit. Hence, a response is
given only when an external go signal is received. The go cluster might
also be activated by endogenous decision processes, but we do not use this
possibility in the following simulations."

For the reciprocal, I reused the term coming from the original paper, it is
to say that speed of learning and single-trial learning are closely related
and for which the aim is to show if the model can learn quickly by making
only one error.

I have already tried to test some lesioning before the code was finished,
and the model was not functioning, I am going to test again to lesion my
model to see what it does. (By lesioning reward and also lesioning memory
of rejected rules).

Yes, auto-evaluation and reasoning are the same thing I think.

I can talk about the first model without modifications, but in fact, there
were also some personal inventions, for the activation of the go unit, as
there was not any explanation in the paper about it.

And I agree to slit the code to see clearer.

This is the first time I am asking for a publication so sorry if I don't do
all things well, thank you for your review, I am going to make the changes,
I don't know if I should answer to your questions by mail or do directly
the changes on the git, tell me if I was right to answer you first.

Best,
Pauline.

Le lun. 9 sept. 2019 à 17:05, Qihong Lu <notifications@github.com> a écrit :

> Wow, this is really interesting! Thanks for the replication. I'm happy to
> help! I will start reading it next week!
>
> I just looked at the code briefly. Just a quick minor comment: Making some
> object-oriented classes to separate the code for defining the model/task
> vs. the code for executing the experiment/plotting the results, might
> further enhance the readability and modularity of code :)
>
