### 2019-09-09 #7 @qihongl
Ah just to be clear I think the code is great!

Actually I'm aware that refactoring the code might be a lot of work and I'm not sure if this is necessary for publishing this work. I was just imagining this work might be used in a cognitive modeling class in the future. I guess having another layer of abstraction can give students the option of forming some overall impression on how the code works first before diving into the detail.
