### 2019-09-18 9468b23
Modifications for 1st review
