### 2019-09-19 #7 @PaulineBock
Thanks for the warning Nicolas. So I have decided to redo the figure and
have stated "adapted from".
Best regards,
Pauline

Le jeu. 19 sept. 2019 à 07:29, Nicolas P. Rougier <notifications@github.com>
a écrit :

> Just step-in as editor-in-chief here: it seems the article includes image
> from the original article and to do that, we need permission from the
> original publisher (e.g. Elsevier, Springer, etc) or you'll need to redo
> the figure (and state "adapted from"). If it is already the case, don't pay
> attention to my message.
>
