### 2019-09-20 #7 @AndreaCaso
Hello,

Thank you for making the changes. I have a few additional notes.

- Reference system should be consistent. The in-text citation for Heaton is (Heaton, R.K, (1981)), and the one for Dehaene and Changeux is [1]. If the reference system you have to abide to is the APA format I'd suggest that you use (Heaton, 1981) and (Dehaene & Changeux, 1991).

- The caption in Fig.3 could potentially be improved, so as to help the reader understand without necessarily referring to the text. I suggest something along these lines: "A comparison between the original model and our re-implementation: Each pattern in the grayscale bar graph represents a different model used  in the original paper (see legend). The bar graphs in colour represent the value obtained with our re-implementation. The overlap in the greyscale and colour bar graphs shows a direct comparison between the value obtained with our re-implementation and those obtained by the original model..."

A few other suggestions:
-  In Fig. 3: 'was succeed' is probably a typo.
- 'The test is no more succeed' could be rephrased.

Provided that the necessary changes are made, I recommend acceptance of the manuscript.

Regards,
Andrea
