### 2019-09-30 #7 @qihongl
btw, the stats code generate plots as the program is being executed. I think it might better to save all plots to some image dir and suppress the output by default?
