### 2019-10-01 #7 @PaulineBock
Hello  @qihongl <https://github.com/qihongl>, thanks for the review.
Concerning the conceptual and discussion questions, I don't have enough
time to work on it, but I have transferred them to my co-author and
previous internship supervisor Frédéric Alexandre.
Concerning the other questions:
-yes we could tell that it is composed of several neural network layers
-the neuronal cluster activities are initially set to 0 before a whole test
of WCST and so before the trials, but not before each block stimuli.
-I will change the criterion explanation, yours is clearer.
-I have added the tree diagram, good idea.
- and the number of execution

- For which figure should I add/change the colors ?

Best,

Pauline.

Le lun. 30 sept. 2019 à 19:00, Qihong Lu <notifications@github.com> a
écrit :

> Thanks! @PaulineBock <https://github.com/PaulineBock> I really enjoyed
> reading it! Here're some thoughts. Please use them as you see fit.
>
> You used a lot of space to describe the order of computations in the
> model, which is good. Though I think the paper can focus more on conceptual
> points such as i) conceptually, what’s the original findings and how does
> their simulation results support that, and ii) what’re the most relevant
> aspects of the original results, or what's the most important patterns of
> capture in terms of a replication, and why your replication qualitatively
> captures that.
>
> I also think more discussion on why are those additional components needed
> to replicate the original results. In the absence of these additional
> mechanisms, how did the original implementation reach the same results? I
> know this is itself a difficult research question... Is it possible to
> contact the original author to get some advice/insights?
>
> The original paper is pretty old. Some terminologies might be unfamiliar
> to many folks. I think connecting them with terminologies can help many
> readers. I was wondering if other folks have the same feeling. For example,
> the phases "neuronal clusters/assemblies” are used in the paper many times.
> I think they are just neural network layers. In this case, pointing our
> that they are the same and stick with one word-choice might be better.
> minor comments/questions:
>
> "Firstly, all neuronal activities are set to 0 before network activation…”
> Do you mean before a trial/or a stimuli block begins?
>
> “The criterion for a change of rule is 3 successes in a row”
> I think some might get confused about this — perhaps explicitly say
> something like "the rule changes whenever the agent made three consecutive
> correct assignment”?
>
> This is optional, but I would summarize the reward contingency as a tree
> diagram (instead of using the five bullet points).
>
> Since the model has many layers, the order of execution is not so easy to
> describe. I would put some numbers on the arrows (in the model figure) to
> make things very clear.
>
> For the new model figure, how about using a different color for the new
> components?
>
> I recommend choosing colors from the colorblind palette:
> https://seaborn.pydata.org/tutorial/color_palettes.html
>
