### 2019-10-01 219c5ee
Saving plots in a directory
No more showing of each plot in a pop-up.
