### 2019-10-03 #7 @qihongl
Thanks for the revision @PaulineBock I think it looks great.

Right, I do think most readers would naturally ask why are there some differences between this implementation versus the original implementation. For example, is it possible that the original paper forgot to report something that's critical for that model to work? Is it possible that there the task implementation / stimuli representation used in the current implementation is slightly different from the original? etc. It would be ideal if there are some responses to this point in the paper. Again, I know this is a challenging point to address. I think contacting the original author to get some feedback might be useful.

Besides that, I don't have additional major concerns. Thanks again!
