### 2019-10-06 #7 @PaulineBock
Hello @qihongl, so we have added some comment about the need of
modifications in the implementation according to the original model. It is
in the Modifiation and Implementation paragraph.

Le jeu. 3 oct. 2019 à 18:38, Qihong Lu <notifications@github.com> a écrit :

> Thanks for the revision @PaulineBock <https://github.com/PaulineBock> I
> think it looks great.
>
> Right, I do think most readers would naturally ask why are there some
> differences between this implementation versus the original implementation.
> For example, is it possible that the original paper forgot to report
> something that's critical for that model to work? Is it possible that there
> the task implementation / stimuli representation used in the current
> implementation is slightly different from the original? etc. It would be
> ideal if there are some responses to this point in the paper. Again, I know
> this is a challenging point to address. I think contacting the original
> author to get some feedback might be useful.
>
> Besides that, I don't have additional major concerns. Thanks again!
>
