### 2019-10-07 #7 @PaulineBock
@oliviaguest I think I have tried to answer to all your questions as I
could. Tell me if it seems okay for you !

Le lun. 7 oct. 2019 à 10:33, Olivia Guest <notifications@github.com> a
écrit :

> What's left to be done here @PaulineBock <https://github.com/PaulineBock>?
> 😄
>
