### 2019-10-07 #7 @oliviaguest
@qihongl @AndreaCaso can you please let me know what you think now with respect to what needs to be done here? For example, are you both proposing acceptance now?
