### 2019-10-07 #7 @qihongl
Sure -- If all of us @PaulineBock  @oliviaguest  @AndreaCaso  are happy with the final version. I think the revision is great.

Just a really minor point: the blank space on page 6 is slightly awkward :) I was wondering if it can be reformatted slightly.
