### 2019-10-08 #7 @AndreaCaso
Hi, I think the revision is very good and I am happy with the final version.
