### 2019-10-08 #7 @PaulineBock
Indeed I have removed the space page 6 !
It is okay for me now.
Thank you all.

Best,
Pauline.

Le mar. 8 oct. 2019 à 10:09, Andrea Caso <notifications@github.com> a
écrit :

> Hi, I think the revision is very good and I am happy with the final
> version.
>
