### 2019-10-15 #7 @oliviaguest
Hey, I'm really sorry... new job and everything is so slow. I'll get on
this tomorrow. :)

Olivia

***
Dr Olivia Guest
sent from my smartphone

On Tue, 15 Oct 2019, 17:10 Nicolas P. Rougier, <notifications@github.com>
wrote:

> Any update ?
>
