### 2019-10-17 #7 @PaulineBock
Hello, I have created my ORCID account. Here is my ID:
https://orcid.org/0000-0002-9374-8733

Thanks for your work.

Le jeu. 17 oct. 2019 à 08:47, Olivia Guest <notifications@github.com> a
écrit :

> @PaulineBock <https://github.com/PaulineBock> do you have an orcid?
> Please make one since you will be asked this so often it is worth just
> doing this now..?
>
