### 2019-10-17 #7 @PaulineBock
I think it is okay ! Thank you really much!
Best,

Pauline.

Le jeu. 17 oct. 2019 à 10:23, Olivia Guest <notifications@github.com> a
écrit :

> @PaulineBock <https://github.com/PaulineBock> thank you ­— as soon as you
> merge PaulineBock/WCSTDehaeneChangeux@master...oliviaguest:patch-1
> <https://github.com/PaulineBock/WCSTDehaeneChangeux/compare/master...oliviaguest:patch-1>
> I'll carry on 😊
>
