### 2019-10-21 #7 @rougier
@oliviaguest Just a little help with the new publication process. @PaulineBock Can you complete your metadata file with the name/ORCID of the editor, reviewers, acceptance data and volume number (=5). As for the issue, I think we can stay on issue 1 or open a new issue (3) because there was a special issue (2). @oliviaguest what do you think?
