### 2019-11-04 #7 @PaulineBock
Hello everyone, sorry for the late answer. So, do someone have all the
informations about the orcid of reviewers and editor ? I will edit as soon
as possible. Thanks.

Le lun. 21 oct. 2019 à 16:03, Olivia Guest <notifications@github.com> a
écrit :

> @rougier <https://github.com/rougier> that sounds fine — thank you!
>
