### 2019-11-04 #7 @PaulineBock
So, I have found @AndreaCaso ORCID ID, but for @qihongl I have found 2
different orcid. What is the date of acceptance ?
Best.

Le lun. 4 nov. 2019 à 14:32, Olivia Guest <notifications@github.com> a
écrit :

> Mine is: 0000-0002-1891-0972
>
