### 2019-11-04 #7 @PaulineBock
Thank you, so metadata is completed for my part.
Tell me if there are some more things to do and when you will publish
please.

Le lun. 4 nov. 2019 à 16:57, Qihong Lu <notifications@github.com> a écrit :

> Sorry about the confusion :)
> here's my orcid: http://orcid.org/0000-0002-0730-5240
>
