### 2019-11-05 #7 @oliviaguest
@rougier what do I need, which directory do I need to be in, in order to run `process.py`? https://github.com/ReScience/articles/blob/master/README.md
