### 2019-11-05 #7 @rougier
Best is to clone the articles repository and to work from the root of that directory (with the PDF and metadata). You can finish filling the metadata and do not forget to use the Zenodo sandbox first to check everythign's fine.

By the way, don't try to publish the website once you've added the new entry into published, just drop me an email or a notification.
