### 2019-11-07 #7 @oliviaguest
@PaulineBock @rougier
I think this is what is needed to be updated in the metadata, correct?

```
Article ID:  416801
Article DOI: 10.5072/zenodo.416801
Article URL: https://sandbox.zenodo.org/record/416801/files/ReScience_Dehaene.pdf
```
