### 2019-11-11 #7 @rougier
Yes the sandbox server is only for you to check everything ran ok.  Since this seems to be the case, you can no run it on the real server and communicate the information to @PaulineBock. Note that the alternative is to modify the metadata yourself and upload the (recompiled) paper and to make a PR to @PaulineBock repo.

Once finished, don't try to "publish" the rescience website since it may fail if you do not have all the plugins. Just add the new bibtex entry to the _published.bib and drop me a notification or an email.
