### 2019-11-18 #7 @oliviaguest
[Sorry, did not meet to close this issue. I keep switching from Mac to Linux which means the shortcuts are confused in my head.]
