### 2019-11-25 #7 @rougier
If you've made the pre-registration on the real Zenodo server, then you or @PaulineBock have to update the metadata file (with DOI, editor name, reviewer name, etc) and regenerate the PDF accordingly (the DOI should now appear on the PDF). Once done, you can register the new PDF at Zenodo still using the script. In case of problem, you still have the option  to manually change the Zenodo entry through the Zenodo website.

Tell me if something's note clear and feel free to improve the description of the process. I realize that I test it for myself knowing the process but in the end the instructions might be a bit confusing.
