### 2019-11-25 #7 @oliviaguest
Does this look right now, @rougier? https://github.com/PaulineBock/WCSTDehaeneChangeux/pull/2
