### 2019-11-25 #7 @rougier
LGTM. Do not forget other informatipon (journal, issue, number, editor, reviewer, etc)
