### 2019-11-26 #7 @oliviaguest
@rougier OK, as soon as the PR is merged. Where do I find the ISSN?

@PaulineBock can you fill in the bits that are for you to do, please?|

```
# Code URL and DOI (url is mandatory for replication, doi after acceptance)
# You can get a DOI for your code from Zenodo,
#   see https://guides.github.com/activities/citable-code/
code:
  - url: https://github.com/PaulineBock/WCSTDehaeneChangeux
  - doi:
```

and:

```
# To be filled by the author(s) after acceptance
# -----------------------------------------------------------------------------

# For example, the URL of the GitHub issue where review actually occured
review:
  - url:
```
