### 2019-11-26 #7 @PaulineBock
Done.

Le mar. 26 nov. 2019 à 11:08, Olivia Guest <notifications@github.com> a
écrit :

> Great, @PaulineBock <https://github.com/PaulineBock> — can you merge the
> PR too, please? PaulineBock/WCSTDehaeneChangeux#2
> <https://github.com/PaulineBock/WCSTDehaeneChangeux/pull/2>
>
