+++ Metadata.yaml
@@ -146,6 +146,6 @@ article:
 # This information will be provided by the editor
 journal:
   - name:   "ReScience C"
-  - issn:   
+  - issn:   ISSN 2430-3658 
   - volume: 5
   - issue:  1
