### 2019-11-28 #7 @oliviaguest
All this info should be on one easy access place, perhaps?

@PaulineBock I think this is the final edit? https://github.com/PaulineBock/WCSTDehaeneChangeux/pull/3
