### 2019-11-28 #7 @PaulineBock
I think yes. Thanks for all your work.

Best.

Le jeu. 28 nov. 2019 à 13:43, Olivia Guest <notifications@github.com> a
écrit :

> All this info should be on one easy access place, perhaps?
>
> @PaulineBock <https://github.com/PaulineBock> I think this is the final
> edit? PaulineBock/WCSTDehaeneChangeux#3
> <https://github.com/PaulineBock/WCSTDehaeneChangeux/pull/3>
>
