### 2019-11-28 #7 @PaulineBock
Is it good now? I have merged the 3 last modifications. (I have marged the
first before seeing that it was a mistake)

Le jeu. 28 nov. 2019 à 14:09, Olivia Guest <notifications@github.com> a
écrit :

> @PaulineBock <https://github.com/PaulineBock> sorry! this is also needed!
> PaulineBock/WCSTDehaeneChangeux#4
> <https://github.com/PaulineBock/WCSTDehaeneChangeux/pull/4>
>
