### 2019-11-28 #7 @oliviaguest
@PaulineBock yes, it looks good — thank you! :smile_cat:

@rougier is everything OK to actually publish to zenodo now?
