### 2019-11-28 #7 @rougier
No, the type of the article is not editorial but replication and the "under review" should not appear anymore in the PDF. There are some comments in the original metadata yaml file to help filling it.
