### 2019-11-28 #7 @oliviaguest
I'm really confused. Type of article is replication? What's wrong with that? https://github.com/PaulineBock/WCSTDehaeneChangeux/blob/11fd0458ceb000329018733d40a59aa9d483e186/Metadata.yaml#L106
