### 2019-11-28 #7 @PaulineBock
For the pdf I must fill Metadata.tex and push pdf again as I work with
overleaf. I am completing the changes.

Le jeu. 28 nov. 2019 à 16:40, Olivia Guest <notifications@github.com> a
écrit :

> I'm really confused. Type of article is replication? What's wrong with
> that?
> https://github.com/PaulineBock/WCSTDehaeneChangeux/blob/11fd0458ceb000329018733d40a59aa9d483e186/Metadata.yaml#L106
>
