### 2019-11-28 #7 @PaulineBock
Yes but I think I had seen that when working on overleaf  the Metadata.tex
cannot be automatically generated from the Metadata.yaml
Not sure but in case I have edited the metadata.tex myself according to the
metadata.yaml.

So now it should be good.

Le jeu. 28 nov. 2019 à 17:10, Olivia Guest <notifications@github.com> a
écrit :

> The type seems correct in the YAML.
>
