### 2019-11-28 #7 @rougier
@PaulineBock There's a python script to convert the metadata.yaml into metadata.tex but else, you can also do it manually.
