### 2019-11-28 #7 @PaulineBock
I have modified it manually. Is it okay for the pdf?

Le jeu. 28 nov. 2019 à 21:18, Nicolas P. Rougier <notifications@github.com>
a écrit :

> @PaulineBock <https://github.com/PaulineBock> There's a python script to
> convert the metadata.yaml into metadata.tex but else, you can also do it
> manually.
>
