### 2019-11-29 #7 @rougier
Not quite yet. I think you're missing the domain (Computational Neuroscience) and you'll need to replace subsection by section (such as to have 1 Introduction instead of 0.1 Introduction)
