### 2019-12-02 #7 @oliviaguest
Edits that need to made (based on @rougier above and me reading through the PDF now):

- [ ] ORCIDs need to be fixed as they currently all point to mine.
- [ ] Domain needs to be Computational Neuroscience.
- [ ] Replace subsection by section (such as to have 1 Introduction instead of 0.1 Introduction).
- [ ] Published date needs to be amended since it didn't get published yet. Would be easy to fix if I could compile your PDF myself so that I can add that in when I publish it.
