### 2019-12-02 #7 @PaulineBock
As I said before I was using overleaf, I can share you my project if you
want. How can I send you a private message ?

Le lun. 2 déc. 2019 à 11:53, Olivia Guest <notifications@github.com> a
écrit :

> Edits that need to made (based on @rougier <https://github.com/rougier>
> above and me reading through the PDF now):
>
>    - ORCIDs need to be fixed as they currently all point to mine.
>    - Domain needs to be Computational Neuroscience.
>    - Replace subsection by section (such as to have 1 Introduction
>    instead of 0.1 Introduction).
>    - Published date needs to be amended since it didn't get published
>    yet. Would be easy to fix if I could compile your PDF myself so that I can
>    add that in when I publish it.
>
