### 2019-12-02 #7 @PaulineBock
I have just edited all that was necessary and pushed on github. If we
publish today it is good, else I will give you the link to edit the
overleaf in private to change the date of publication.

Le lun. 2 déc. 2019 à 13:17, Bock Pauline <mllebock@gmail.com> a écrit :

> As I said before I was using overleaf, I can share you my project if you
> want. How can I send you a private message ?
>
> Le lun. 2 déc. 2019 à 11:53, Olivia Guest <notifications@github.com> a
> écrit :
>
>> Edits that need to made (based on @rougier <https://github.com/rougier>
>> above and me reading through the PDF now):
>>
>>    - ORCIDs need to be fixed as they currently all point to mine.
>>    - Domain needs to be Computational Neuroscience.
>>    - Replace subsection by section (such as to have 1 Introduction
>>    instead of 0.1 Introduction).
>>    - Published date needs to be amended since it didn't get published
>>    yet. Would be easy to fix if I could compile your PDF myself so that I can
>>    add that in when I publish it.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/ReScience/submissions/issues/7?email_source=notifications&email_token=AIC3KE7KARZD2ZYU234X22DQWTSI7A5CNFSM4ISBMHCKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEFTCSPY#issuecomment-560343359>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AIC3KEZDNAX2QMRJGMZY6ZTQWTSI7ANCNFSM4ISBMHCA>
>> .
>>
>
