### 2019-12-04 #7 @oliviaguest
This is how: https://help.github.com/en/github/setting-up-and-managing-your-github-user-account/inviting-collaborators-to-a-personal-repository
