### 2019-12-04 #7 @PaulineBock
Thanks for the tips, it is done :)

Le mer. 4 déc. 2019 à 13:04, Olivia Guest <notifications@github.com> a
écrit :

> This is how:
> https://help.github.com/en/github/setting-up-and-managing-your-github-user-account/inviting-collaborators-to-a-personal-repository
>
