### 2019-12-04 #7 @oliviaguest
Sorry for the delay, just working through some stuff @PaulineBock!

@rougier when you get a chance please feel free to advise... :sweat_smile:  https://github.com/ReScience/articles/issues/2
