### 2019-12-09 #7 @PaulineBock
Thank to all of you!! It was a pleasure to work with you :)

Le lun. 9 déc. 2019 à 15:41, Olivia Guest <notifications@github.com> a
écrit :

> Closed #7 <https://github.com/ReScience/submissions/issues/7>.
>
